+++ Src/Transform.h
@@ -4,7 +4,7 @@ struct TransformConfig
 {
 	int x;
 	int y;
-	int angle;
+	double angle;
 	double scale;
 };
 
@@ -12,13 +12,13 @@ class Transform
 {
 public:
 	Transform(class Object* obj, int x = 0, int y = 0,
-		int angle = 0, double scale = 1);
+		double angle = 0, double scale = 1);
 
 	void MoveTo(int x, int y);
 	void Move(int deltaX, int deltaY);
 
-	void RotateTo(int angle);	// Use Angle System
-	void Rotate(int angle);
+	void RotateTo(double angle);	// Use Angle System
+	void Rotate(double angle);
 
 	void ScaleTo(double scale);
 	void Scale(double scale);
@@ -27,8 +27,8 @@ public:
 
 	int GetX();
 	int GetY();
-	int GetAngle();
-	int GetScale();
+	double GetAngle();
+	double GetScale();
 
 	TransformConfig GetTransform();
 
@@ -38,6 +38,6 @@ private:
 	int x;
 	int y;
 
-	int angle;
+	double angle;
 	double scale;
 };
